+++ NEWS.md
@@ -1,3 +1,9 @@
 # Rsurveytools 0.1.0
+* Initial release of Rsurveytools
 
-* Initial CRAN submission.
+
+# Rsurveytools 0.2.0
+
+## New features
+* Renamed original `crosstab()` function to `srvy_crosstab()` for survey objects
+* Added new `crosstab()` function for standard data frames